# api/src/main/resources/public/api.yaml
openapi: 3.0.1
info:
  title: Car Hire API
  version: 0.1.0
  contact:
    name: Dragomir Alin
  description: Car Hire API
servers:
  - url: http://localhost:8080
    description: local environment
    variables:
      port:
        default: '8080'
...
paths:
  /user:
    get:
      tags:
        - user
      summary: get all users
      description: returns a list of all the users, without any filtering
      operationId: getUsers
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponseDTO'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
    post:
      tags:
        - user
      summary: add a new user
      description: adds a new user and returns it
      operationId: addUser
      requestBody:
        description: The data of the user to insert.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDTO'
            example:
              email: alin@carhire.com
              password: mypa$$w0rd!@$
              username: dragomiralin
              firstname: Dragomir
              title: Employee
              lastname: Alin
              countryofresidence: Romania
              driverslicensenr: 221004
              status: ACTIVE
      responses:
        '201':
          $ref: '#/components/responses/Created'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
  /user/{id}:
    get:
      tags:
        - user
      summary: get user
      description: returns the details for a user, based on id
      operationId: getUser
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponseDTO'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
    patch:
      tags:
        - user
      summary: update user
      description: update user
      operationId: updateUser
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        $ref: '#/components/requestBodies/PatchBody'
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDTO'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
    delete:
      tags:
        - user
      summary: remove a user
      description: delete a user identified by id
      operationId: removeUser
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: success
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
...
components:
  responses:
    Created:
      description: created with success
      content:
        application/json:
          schema:
            type: string
            properties:
              id:
                type: string
                example: 12fc2
    NotFound:
      description: not found
    Unauthorized:
      description: unauthorized
    BadRequest:
      description: bad request
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: '#/components/schemas/ValidationDTO'
    ServerError:
      description: internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorDTO'
  requestBodies:
    PatchBody:
      description: A message in JsonPatch format
      required: true
      content:
        application/json-patch+json:
          schema:
            $ref: '#/components/schemas/PatchBody'
  schemas:
    PatchBody:
      type: array
      items:
        $ref: '#/components/schemas/PatchDocument'
    PatchDocument:
      type: object
      description: A JSONPatch document as defined by RFC 6902.
      required:
        - 'op'
        - 'path'
        - 'value'
      properties:
        op:
          type: string
          description: The operation to be performed
          enum:
            - 'add'
            - 'remove'
            - 'replace'
            - 'move'
            - 'copy'
            - 'test'
        path:
          type: string
          description: A JSON-Pointer
        value:
          type: object
          description: The value to be used within the operations.
        from:
          type: string
          description: A string containing a JSON Pointer value.
    ValidationDTO:
      type: object
      properties:
        code:
          type: string
        field:
          type: string
        description:
          type: string
    ErrorDTO:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
    UserResponseDTO:
      type: object
      properties:
        id:
          type: string
          description: id of the user
        email:
          type: string
          description: email of the user
        username:
          type: string
          description: username of the user
        firstname:
          type: string
          description: first name of the user
        lastname:
          type: string
          description: last name of the user
        countryofresidence:
          type: string
          description: country of residence of the user
        title:
          type: string
          description: title of the user
        driverslicensenr:
          type: integer
          format: int32
          description: driver`s license number of the user
        status:
          type: string
          enum: [ ACTIVE, CANCELLED ]
          description: the status of the user
    UserDTO:
      type: object
      properties:
        id:
          type: integer
          format: int32
          description: id of the user
        email:
          type: string
          description: email of the user
        password:
          type: string
          format: password
          description: password of the user
        username:
          type: string
          description: username of the user
        firstname:
          type: string
          description: first name of the user
        lastname:
          type: string
          description: last name of the user
        countryofresidence:
          type: string
          description: country of residence of the user
        title:
          type: string
          description: title of the user
        driverslicensenr:
          type: integer
          format: int32
          description: driver`s license number of the user
        status:
          type: string
          enum: [ACTIVE, CANCELLED]
          description: the status of the user
...